feat: split reth and hera binaries
