+++ bin/reth/README.md
@@ -1,0 +1,1 @@
+# `reth`
