+++ crates/hera-exex/src/lib.rs
@@ -1,0 +1,8 @@
+mod exex;
+pub use exex::HeraExEx;
+
+mod cli;
+pub use cli::HeraArgsExt;
+
+/// The identifier of the Hera Execution Extension.
+pub const HERA_EXEX_ID: &str = "hera";
